requirements.txt
yt-dlp
python-dotenv
google-api-python-client
pandas
openpyxl